+++ web_v2/requests/base.go
@@ -1,9 +1,11 @@
 package requests
 
 import (
+	"github.com/spf13/cast"
 	"net/http"
 	"net/url"
 	"reflect"
+	"regexp"
 	"strconv"
 	"strings"
 
@@ -115,16 +117,33 @@ var amountHook = figure.Hooks{
 	},
 }
 
+func mkJsonTag(fieldName string) string {
+	return fmt.Sprintf("json:\"%s\"", fieldName)
+}
+
+func toSnakeCase(str string) string {
+	matchFirstCap := regexp.MustCompile("(.)([A-Z][a-z]+)")
+	matchAllCap := regexp.MustCompile("([a-z0-9])([A-Z])")
+
+	snake := matchFirstCap.ReplaceAllString(str, "${1}_${2}")
+	snake = matchAllCap.ReplaceAllString(snake, "${1}_${2}")
+
+	return strings.ToLower(snake)
+}
+
 func (r *base) populateFilters(target interface{}) error {
+	//_ = r.validateFilters(target)
 	filter := make(map[string]interface{})
 	for k, v := range r.filter {
 		filter[k] = v
 	}
 
 	err := figure.Out(target).With(figure.BaseHooks, amountHook).From(filter).Please()
-
 	if err != nil {
-		return err
+		f := errors.GetFields(err)
+		return validation.Errors{
+			toSnakeCase(cast.ToString(f["field"])): err,
+		}
 	}
 
 	return nil