+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * panic on handle set fee operation when account or account role does not exist
 * panic on actions' `isAllowed` checking
 * error on get reviewable request by reviewer
+* order book is now sorted
+* create change role ingestion causing 500 on operations
 
 ### Removed