+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `GetBalanceList` handler to handle `v3/balances` request
 * `GetPublicKeyEntry` handler to handle `v3/public_key_entries/{id}` request
 * Receiver filter for create issuance requests
+* Now Bad Request is correctly returned in case of invalid signature
 
 ### Fixed